package entity renamed to entities, changed Lists to Sets
