+++ src/main/java/cz/fi/muni/pa165/enums/Taste.java
@@ -1,5 +1,5 @@
 package cz.fi.muni.pa165.enums;
 
 public enum Taste {
-    SWEET, SEMI_SWEET, SEMI_SOUR, SOUR
+    SWEET, SEMI_SWEET, SEMI_DRY, DRY
 }
